Answer Based Question Of BS where i solved another type which is based on min(max) & max(min) and MedianTwoSortedArray Using Binary Search.

+++ src/BinarySearch/MinimizeMaxDistanceToGasStation.java
@@ -16,7 +16,7 @@ public class MinimizeMaxDistanceToGasStation {
 
         //Find the maximum distance:
         for (int i = 0; i < n - 1; i++) {
-            high = Math.max(high, arr[i + 1] - arr[i]);
+            high = Math.max(high, (double)(arr[i + 1] - arr[i]));
         }
 
         //Apply Binary search:
